requirements.txt
Flask>=2.3
Flask-WTF>=1.1
pydantic>=2.0
Werkzeug>=3.0
Flask-SQLAlchemy>=3.0
Flask-Migrate>=4.0
pytest
email-validator
